refactor(frontend): run landing prerender via vite ssr build step
